+++ cmd/juno/verify/verify.go
@@ -51,69 +51,14 @@ func verifyAll(cmd *cobra.Command, args []string) error {
 
 	trieVerifier := NewTrieVerifier(database, logger)
 
-	runner := &VerifyRunner{
+	verifier := &VerifyRunner{
 		Verifiers: []Verifier{
 			trieVerifier,
 		},
 	}
 
-	configs := make(map[string]Config)
 	ctx := cmd.Context()
-	return runner.Run(ctx, configs)
-}
-
-func verifyTrieCmd() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "trie",
-		Short: "Verify trie integrity",
-		Long:  `Verify trie integrity by rebuilding tries from leaf nodes and comparing root hashes.`,
-		RunE:  runTrieVerify,
-	}
-
-	cmd.Flags().StringSlice(
-		verifyTrieType,
-		nil,
-		"Trie types to verify (state, class, contract). Can be specified multiple times. Empty = all.",
-	)
-
-	return cmd
-}
-
-func runTrieVerify(cmd *cobra.Command, args []string) error {
-	dbPath, err := cmd.Flags().GetString(verifyDBPathF)
-	if err != nil {
-		return err
-	}
-
-	database, err := openDB(dbPath)
-	if err != nil {
-		return err
-	}
-	defer database.Close()
-
-	trieTypes, err := cmd.Flags().GetStringSlice(verifyTrieType)
-	if err != nil {
-		return err
-	}
-
-	cfg := &TrieConfig{}
-
-	if len(trieTypes) > 0 {
-		cfg.Tries = make([]TrieType, len(trieTypes))
-		for i, t := range trieTypes {
-			cfg.Tries[i] = TrieType(t)
-		}
-	}
-
-	logLevel := utils.NewLogLevel(utils.INFO)
-	logger, err := utils.NewZapLogger(logLevel, true)
-	if err != nil {
-		return fmt.Errorf("failed to create logger: %w", err)
-	}
-
-	verifier := NewTrieVerifier(database, logger)
-	ctx := cmd.Context()
-	return verifier.Run(ctx, cfg)
+	return verifier.Run(ctx, nil)
 }
 
 func openDB(path string) (db.KeyValueStore, error) {
